pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "vista-docs"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "requests",
    "python-docx",
    "pyyaml",
    "click",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "pytest-watch",
    "pytest-randomly",
    "ruff",
    "mypy",
    "pre-commit",
    "types-requests",
    "types-PyYAML",
]

[project.scripts]
vista-docs = "vista_docs.cli.main:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short -x"
markers = [
    "network: marks tests that require live VA network access (deselect with '-m not network')",
    "slow: marks tests that are slow to run",
    "integration: marks integration tests",
]

[tool.coverage.run]
source = ["src"]

[tool.coverage.report]
fail_under = 80
show_missing = true

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
